docs: Waku router under the hood
This is an idea for a router post. I am happy to reword, restructure
or whatever.

It was helpful for me to organize my thoughts to look a bit
closer at what is loaded when on each route, so I think a
post to talk through some of this has some value.

Excited to hear what y'all think :)