docs: Waku router under the hood and add guides to website (#1311)
This is an idea for a router post. I am happy to reword, restructure or
whatever.

It was helpful for me to organize my thoughts to look a bit closer at
what is loaded when on each route, so I think a post to talk through
some of this has some value.

Excited to hear what y'all think :)

---------

Co-authored-by: Tyler <26290074+thegitduck@users.noreply.github.com>