Support installing Waku package from git repo
Prior to this commit, when installing Waku from a git repository, the
`dist` directory would not be built, and all `waku` CLI commands would
throw an `ERR_MODULE_NOT_FOUND` error.

This commit adds a `prepare` command to `packages/waku/package.json`,
which is just an alias for `pnpm run compile`.  pnpm will automatically
run `prepare` when installing the package from a git repo, thus solving
the problem.

Also, documentation about referencing a package inside a monorepo seems
to be scarce and inaccurate (many sources specify incorrect syntax).
Therefore, this commit adds instructions to the README on how to build
an app that uses an experimental version of Waku.