fix(cli/dev only): workaround to ignore "TypeError [ERR_INVALID_STATE]: Invalid state: Unable to enqueue" (#1407)
This PR introduces a workaround for issue #756.
I've also been facing the same issue often.

# What's happening

Please look at 00:35 to observe the error.
I just added new lines to the file, then restarted the process.


https://github.com/user-attachments/assets/bf6b038c-10b9-435a-b85c-97651fc7b24e

# Workaround

It seems that the error comes from TransformStream, but all of the code
seems to have no issue, and it's too hard to investigate.
However, as you can see in the above video, the error might be
ignorable.
My suggestion is to catch the error and ignore it in only the dev CLI.


https://github.com/user-attachments/assets/4964197e-7594-49ab-a71f-c8398f25e168

You can see the result at 00:42.

Co-authored-by: Daishi Kato <dai-shi@users.noreply.github.com>