fix: continue transition with waku/router Link (#1458)
Part of the fix for #1437.

It turns out that #1437 is reporting multiple issues:

1. Showing preview element content instead of a transition when
navigating to a previously visited page. That issue is replicated with
the e2e tests in this PR and fixed.
2. Updating the url before changing the route. I'm expecting the url to
not update until the new page has loaded. I opened #1459.
3. The changeRoute function completes before the page component has
streamed any data.

The third issue is hard to test without using unstable_events.

---------

Co-authored-by: daishi <daishi@axlight.com>