fix(create-pages): Fix API catch-all routes matching before standard API routes (#1508)
### Details:

This change adds e2e test cases for both issues:

- catch-all api route incorrectly matching before standard api route
#1507
- ~~catch-all route not handling index route~~ -> moved this to it's own
issue: #1518

and also adds a fix to the first issue above by sorting the routes first
and then finding the first match. The sorting logic now accounts for
"specificity" of the route (e.g. number of path specs) and also if the
path spec is a wildcard.