fix: auto optimize `use-sync-external-store` (#1594)
- Using https://github.com/vitejs/vite-plugin-react/pull/674

I'm thinking to add a slight hack on plugin side to support
`use-sync-external-store` out-of-the-box, which tends to be smuggled
within deep dependency tree and causes a cryptic error. After this
change, `e2e/fixtures/ssr-swr/waku.config.ts` can be removed.

I thought this would just work without change on Waku, but it failed
because it matches `@vitejs/plugin-rsc/` instead of
`@vitejs/plugin-rsc`.