perf: reuse RSC render in HTML render during SSG (#1765)
Currently each page server component is evaluated twice for `renderRsc`
and `renderHtml`. In principle, it looks safe to reuse `renderRsc`'s
output stream into `renderHtml` as pre-rendered rsc stream by `tee`-ing
stream, which would reduce server component evaluation.

This idea is based on SSG example of mine in
https://github.com/vitejs/vite-plugin-react/blob/a06940b10c3d7c19236f1f176fc0a1748b9f62e7/packages/plugin-rsc/examples/ssg/src/framework/entry.rsc.tsx#L41-L59.
I'm curious if this approach is attempted in the past and if there's a
known issue with the idea.

TODO
- [x] test
- we can verify by `performance.now()` appearing exact same time both in
ssr html and rsc (for client side navigation)
- [x] bench
  - `packages/website`
  - `e2e/fixtures/ssg-performance`

---

- `packages/website`

```sh
# before
[ssg] processing static generation...
✓ finished in 2271ms

# after
[ssg] processing static generation...
✓ finished in 1491ms
```

- `e2e/fixtures/ssg-performance` (this only stress-tests number of pages
without CPU work, so it's less obvious)

```sh
# before
[ssg] processing static generation...
✓ finished in 4997ms

# after
[ssg] processing static generation...
✓ finished in 4599ms
```

---------

Co-authored-by: Daishi Kato <dai-shi@users.noreply.github.com>