feat: export vite plugins (unstable) (#1864)
This PR proposes to export `rscPlugin` via the internal entrypoint. I
intentionally did not add it in `src/internals.ts` as it would conflict
with the virtual modules exposed in `createServerEntryAdapter`.

Rationale: Exploring a new version of
[Vocs](https://github.com/wevm/vocs) that will use Waku internally (in a
Vite plugin). As such, it would be amazing to have access to the
`rscPlugin`!

----

edit(daishi): export all internal plugins instead of `rscPlugin`.

---------

Co-authored-by: daishi <daishi@axlight.com>